+++ src/lv_core/lv_obj.h
@@ -838,14 +838,6 @@ lv_res_t lv_event_send_refresh(lv_obj_t * obj);
 void lv_event_send_refresh_recursive(lv_obj_t * obj);
 
 /**
- * Queue the sending of LV_EVENT_REFRESH event to an object and all of its children.
- * The events won't be sent immediately but after `LV_DISP_DEF_REFR_PERIOD` delay.
- * It is useful to refresh object only on a reasonable rate if this function is called very often.
- * @param obj pointer to an object or NULL to refresh all objects of all displays
- */
-void lv_event_queue_refresh_recursive(lv_obj_t * obj);
-
-/**
  * Call an event function with an object, event, and data.
  * @param event_xcb an event callback function. If `NULL` `LV_RES_OK` will return without any actions.
  *        (the 'x' in the argument name indicates that its not a fully generic function because it not follows